requirements.txt
anthropic
openai
watchdog
requests
Pillow
flask
